// src/types/api.ts
export interface ImageResponse {
  id: string;
  imageUrl: string;
  publicId?: string;
}
// ...
export interface CertificationRequest {
  title: string;
  issuingAuthority: string;
  certificateNumber: string;
  issuedAt: string;
  expireAt: string;
  description: string;
}

export interface CertificationResponse {
  id: string;
  title: string;
  issuingAuthority: string;
  certificateNumber: string;
  issuedAt: string;
  expireAt: string;
  description: string;
  certificateImages: ImageResponse[];
}